fix sdk package not available
