+++ sdk/typescript/rooch-sdk/src/address/address.ts
@@ -2,9 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { Bytes } from '../types/index.js'
-import { BitcoinAddress, BitcoinNetowkType } from './bitcoin.js'
-import { NoStrAddress } from './nostr.js'
-import { RoochAddress } from './rooch.js'
 
 export const ROOCH_BECH32_PREFIX = 'rooch'
 
@@ -13,23 +10,4 @@ export const ROOCH_ADDRESS_LENGTH = 32
 export interface Address {
   toBytes(): Bytes
   toStr(): string
-}
-
-export class AddressView implements Address {
-  public readonly bitcoinAddress: BitcoinAddress
-  public readonly noStrAddress: NoStrAddress
-  public readonly roochAddress: RoochAddress
-
-  constructor(publicKey: Bytes, network: BitcoinNetowkType = BitcoinNetowkType.Regtest) {
-    this.bitcoinAddress = BitcoinAddress.fromPublicKey(publicKey, network)
-    this.noStrAddress = new NoStrAddress(publicKey)
-    this.roochAddress = this.bitcoinAddress.genRoochAddress()
-  }
-
-  toBytes(): Uint8Array {
-    return this.roochAddress.toBytes()
-  }
-  toStr(): string {
-    return this.roochAddress.toStr()
-  }
 }
