+++ sdk/typescript/rooch-sdk/src/address/index.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from './address.js'
+export * from './addressView.js'
 export * from './rooch.js'
 export * from './bitcoin.js'
 export * from './util.js'
